+++ changelog.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - [Add `@pattern` constraint on string types](https://github.com/ballerina-platform/ballerina-standard-library/issues/3179)
+- [Add constraint support for readonly types](https://github.com/ballerina-platform/ballerina-standard-library/issues/3742)
 
 ## [1.0.1] - 2022-11-29