+++ changelog.md
@@ -3,6 +3,12 @@ This file contains all the notable changes done to the Ballerina Constraint pack
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- [Ensure value type after constraint validation](https://github.com/ballerina-platform/ballerina-standard-library/issues/3976)
+
 ## [1.0.2] - 2023-01-30
 
 ### Fixed