+++ changelog.md
@@ -3,7 +3,7 @@ This file contains all the notable changes done to the Ballerina Constraint pack
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2023-02-20
 
 ### Fixed
 - [Fix unexpected errors due to custom annotations](https://github.com/ballerina-platform/ballerina-standard-library/issues/3817)